backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.135.0",
    "python-multipart<1.0.0,>=0.0.7",
    "email-validator<3.0.0.0,>=2.1.0.post1",
    "tenacity<9.0.0,>=8.2.3",
    "pydantic>2.0",
    "emails<1.0,>=0.6",
    "jinja2<4.0.0,>=3.1.4",
    "alembic<2.0.0,>=1.12.1",
    "httpx>=0.25.1,<1.0.0",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "sqlmodel<1.0.0,>=0.0.21",
    "pydantic-settings<3.0.0,>=2.2.1",
    "sentry-sdk[fastapi]>=2.20.0,<3.0.0",
    "pyjwt<3.0.0,>=2.8.0",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "proxmoxer>=2.2.0",
    "websockets>=16.0",
    "cryptography>=43.0.0",
    "paramiko>=3.4.0",
    "redis[hiredis]>=5.0.0",
    "pyyaml>=6.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "prometheus-client>=0.20.0",
]

[dependency-groups]
dev = [
    "pytest<8.0.0,>=7.4.3",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "prek>=0.2.24,<1.0.0",
    "coverage<8.0.0,>=7.4.3",
]

[tool.pytest.ini_options]
# Auto-mode: async test functions are awaited without needing
# @pytest.mark.asyncio decorator on each one. Existing decorators stay valid.
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.ruff.lint.per-file-ignores]
# Re-export modules legitimately import-and-re-expose names.
"__init__.py" = ["F401"]
# FastAPI route handlers receive auth/permission dependencies (e.g.
# `current_user: AdminUser`) that act as guards even when the body never
# reads them. Same pattern applies to lifespan/middleware factories and
# WebSocket / rate-limit dependency callables.
"app/api/routes/*" = ["ARG001"]
"app/api/deps/*" = ["ARG001"]
"app/main.py" = ["ARG001"]
# Service modules in active development — interface conformance and
# placeholder parameters intentionally remain unused while behind feature
# flags or future refactors. Do not silently fix.
"app/services/vm/*" = ["ARG001", "F841"]
"app/services/scheduling/*" = ["ARG001"]
"app/services/network/*" = ["ARG001"]
# Tests rely on print() for debug output and accept fixture/mock arguments
# that are intentionally unused. Some env-bootstrap tests must set env vars
# before the import they exercise (E402).
"tests/*" = ["T201", "ARG001", "F401", "E402", "F841", "C408"]

[tool.coverage.run]
source = ["app"]
dynamic_context = "test_function"

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true
